fix(ci): 修复多个编译错误解决CI构建失败
- 修复MainActivity.kt缺少mutableStateOf和Column导入
- 修复PlayerScreen.kt时间参数类型从Int改为Long
- 移除DatabaseModule.kt中未使用的Hilt/Dagger注解
- 修复MusicService.kt中MediaSession Token类型不匹配
- 修复HomeScreen.kt中tabIndicatorOffset在Material3中已移除
- 修复Logger.kt中Array不支持removeAt
- 修复ViewModel中const val不能在类实例中使用
- 更新版本号到0.2.2

+++ app/src/main/java/app/music_q36_qwen_code/ui/screens/PlayerScreen.kt
@@ -14,6 +14,7 @@ import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.*
 import androidx.compose.material3.*
 import androidx.compose.runtime.*
+import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.blur
@@ -39,13 +40,13 @@ import kotlin.math.abs
 fun PlayerScreen(
     currentSong: Song?,
     isPlaying: Boolean,
-    currentPosition: Int,
-    duration: Int,
+    currentPosition: Long,
+    duration: Long,
     lyrics: List<Pair<Long, String>>,
     onPlayPauseClick: () -> Unit,
     onNextClick: () -> Unit,
     onPreviousClick: () -> Unit,
-    onSeekTo: (Int) -> Unit,
+    onSeekTo: (Long) -> Unit,
     onBackClick: () -> Unit,
     onFavoriteClick: () -> Unit,
     isFavorite: Boolean
@@ -239,13 +240,13 @@ fun TopBar(
 fun CoverModeContent(
     currentSong: Song,
     isPlaying: Boolean,
-    currentPosition: Int,
-    duration: Int,
+    currentPosition: Long,
+    duration: Long,
     lyrics: List<Pair<Long, String>>,
     onPlayPauseClick: () -> Unit,
     onNextClick: () -> Unit,
     onPreviousClick: () -> Unit,
-    onSeekTo: (Int) -> Unit
+    onSeekTo: (Long) -> Unit
 ) {
     Column(
         horizontalAlignment = Alignment.CenterHorizontally,
@@ -338,7 +339,7 @@ fun LyricsModeContent(
     songTitle: String,
     artist: String,
     lyrics: List<Pair<Long, String>>,
-    currentPosition: Int,
+    currentPosition: Long,
     onPlayPauseClick: () -> Unit,
     isPlaying: Boolean
 ) {
@@ -448,9 +449,9 @@ fun LyricsModeContent(
 
 @Composable
 fun ProgressBar(
-    currentPosition: Int,
-    duration: Int,
-    onSeekTo: (Int) -> Unit
+    currentPosition: Long,
+    duration: Long,
+    onSeekTo: (Long) -> Unit
 ) {
     val currentProgress = if (duration > 0) currentPosition.toFloat() / duration else 0f
     var sliderPosition by remember { mutableFloatStateOf(currentProgress) }
@@ -593,7 +594,7 @@ fun ModeButton(
     }
 }
 
-fun formatTime(milliseconds: Int): String {
+fun formatTime(milliseconds: Long): String {
     val seconds = milliseconds / 1000
     val minutes = seconds / 60
     val remainingSeconds = seconds % 60
